Don't parse in memory pem files in reverse order
Downstream consumers of readSignedObjectFromMemory rely on the order of
certificates parsed from a pem file. Return them in the same order that
they are in memory. This matches the results of readSignedObject.

Change the parsing of keys to be the same order as they are on the disk
as well. The overwhelming number of key files will have one key in them.
For those, this will be a no-op. In the case that there are more keys,
the downstream consumer is almost certainly expecting them in the order
they are on disk.